backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
aiosqlite>=0.20.0
alembic>=1.14.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.18
pydantic>=2.10.0
pydantic-settings>=2.7.0
email-validator>=2.1.0
aiofiles>=24.1.0
python-dotenv>=1.0.1
httpx>=0.28.0
beautifulsoup4>=4.12.3
Jinja2>=3.1.4
websockets>=14.1
boto3>=1.35.0
pypdf>=5.0.0
python-docx>=1.1.0
